requirements.txt
tensorflow-macos
pandas
numpy
matplotlib
streamlit
seaborn
scikit-learn
h5py
